Замена int на double

+++ МиСПИСиТ, лаба1/Form1.cs
@@ -20,15 +20,15 @@ namespace МиСПИСиТ__лаба1
 
         private void plus_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(tBa.Text);
-            int b = Convert.ToInt32(tBb.Text);
+            double a = Convert.ToDouble(tBa.Text);
+            double b = Convert.ToDouble(tBb.Text);
             tB.Text = Convert.ToString(a + b);
         }
 
         private void minus_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(tBa.Text);
-            int b = Convert.ToInt32(tBb.Text);
+            double a = Convert.ToDouble(tBa.Text);
+            double b = Convert.ToDouble(tBb.Text);
             tB.Text = Convert.ToString(a - b);
         }
     }
